btcusd.core docker yml

+++ Core.App/ApiModels/BitFinex/BitFinexModel.cs
@@ -11,6 +11,6 @@ namespace Core.App.ApiModels.BitFinex
         public decimal? low { get; set; }
         public decimal? high { get; set; }
         public decimal? volume { get; set; }
-        public int? timestamp { get; set; }
+        public long? timestamp { get; set; }
     }
 }
